Combine duplicated code.
